+++ src/Flag.js
@@ -1,7 +1,5 @@
-import _ from "lodash";
-import React from "react";
-
-const PropTypes = React.PropTypes;
+import React, {PropTypes} from "react";
+import Flag from "./Flag.js";
 
 export default React.createClass({
   /**
@@ -12,11 +10,22 @@ export default React.createClass({
     // Name of country or region for this flag.
     name: PropTypes.string,
 
+    format: PropTypes.oneOf(["png", "icns", "ico"]),
+
     // Size of the country flag
-    size: PropTypes.oneOf(["16", "24", "32", "48", "64", "icns", "ico"]),
+    pngSize: PropTypes.oneOf([16, 24, 32, 48, 64]),
 
     // Size of the country flag
     shiny: PropTypes.bool,
+
+    // Width of the flag <img> HTML tag.
+    width: PropTypes.number,
+
+    // Height of the flag <img> HTML tag.
+    height: PropTypes.number,
+
+    // Alternative text of the flag <img> HTML tag.
+    alt: PropTypes.string,
   },
 
   /**
@@ -28,17 +37,20 @@ export default React.createClass({
       // Default flag.
       name: "_unknown",
 
+      // Default format in PNG.
+      format: "png",
+
       // Default size of 32 pixels.
-      size: "32",
+      pngSize: 32,
 
       // Shiny flags (Default is flat)
-      shiny: false
-    };
-  },
+      shiny: false,
 
-  getInitialState() {
-    return {
+      width: null,
 
+      height: null,
+
+      alt: null
     };
   },
 
@@ -47,27 +59,29 @@ export default React.createClass({
    */
 
   render() {
-    let type = this.props.shiny ? "shiny" : "flat";
+    const type = this.props.shiny ? "shiny" : "flat";
 
-    let extension = _.isNaN(parseInt(this.props.size)) ?
-      this.props.size
-    : "png";
+    const folder = (this.props.format === "icns" || this.props.format === "ico") ?
+      this.props.format : this.props.pngSize;
+
+    const altText = this.props.alt ? this.props.alt : this.props.name;
 
     return (
-      <span>
-        <img
-          src={
-            "../vendor/flags/flags-iso/" +
-            type +
-            "/" +
-            this.props.size +
-            "/" +
-            this.props.name +
-            "." +
-            extension
-          }
-        />
-      </span>
+      <img
+        alt={altText}
+        src={
+          "../vendor/flags/flags-iso/" +
+          type +
+          "/" +
+          folder +
+          "/" +
+          this.props.name +
+          "." +
+          this.props.format
+        }
+        width={this.props.width}
+        height={this.props.height}
+      />
     );
   }
 });