+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## v0.1.1 - (2015/06/17)
+
+* Possiblity to set the <img> HTML tag: width, height and alt text in the props.
+* Changed props structure (file type vs size).
+
 ## v0.1.0 - (2015/06/17)
 
 * Initial version of the project.